# backend/internal/api/server.go
// Package api wires HTTP routing, middleware and handlers.
package api

import (
	"context"
	"log/slog"
	"net"
	"net/http"
	"strings"
	"sync"
	"time"

	"github.com/3219378872/rent-auto/backend/internal/auth"
	"github.com/3219378872/rent-auto/backend/internal/domain"
	"github.com/3219378872/rent-auto/backend/internal/platform/uu"
	"github.com/3219378872/rent-auto/backend/internal/store"
)

const timeRFC3339Milli = "2006-01-02T15:04:05.000Z07:00"

type Server struct {
	Store             *store.Store
	JWT               *auth.JWT
	TTL               time.Duration
	AdminUser         string
	Version           string
	Log               *slog.Logger
	Jobs              JobController   // nil-safe: endpoints degrade when nil
	Channels          ChannelsService // nil-safe
	Wallets           WalletProvider  // nil-safe
	Steam             SteamService    // nil-safe
	EnvironmentDryRun *bool           // nil means execution mode is not configured/unknown
	// PasswordHash resolves the current admin bcrypt hash (env- or DB-backed).
	PasswordHash   func(ctx context.Context) (string, error)
	PasswordChange func(ctx context.Context, expectedHash, newHash string) error
	Epochs         EpochStore

	logins       *loginLimiter
	smsMu        sync.Mutex
	smsFails     map[string]*failSlot // per-IP UU sms send throttle
	dummyOnce    sync.Once
	dummyHashVal string
	trustMu      sync.RWMutex
	trustProxies []*net.IPNet // explicit TRUST_PROXY_CIDRS; empty = default private ranges
	defaultTrust []*net.IPNet // lazily parsed defaultTrustCIDRs
}

type EpochStore interface {
	SessionEpoch(context.Context) (int64, error)
	BumpSessionEpoch(context.Context) (int64, error)
}

type SteamService interface {
	Health(ctx context.Context) string
	SetCredentials(ctx context.Context, username, password, sharedSecret, identitySecret string) error
}

type ChannelsService interface {
	Health(ctx context.Context) map[string]string
	SendLoginSmsCode(ctx context.Context, phone, sessionID string, captcha *uu.CaptchaResult) (uu.SmsCodeResult, error)
	GetSmsUpSignInConfig(ctx context.Context) (uu.SmsUpConfig, error)
	VerifyUUSms(ctx context.Context, phone, code, sessionID, loginReqTicket string) (string, error)
	SetUUToken(ctx context.Context, token string) error
	SetECOCreds(ctx context.Context, partnerID, privateKeyPEM, steamID string) error
}

func NewServer(st *store.Store, jwt *auth.JWT, adminUser, version string, log *slog.Logger) *Server {
	return NewServerWithTTL(st, jwt, adminUser, version, 24*time.Hour, log)
}

// NewServerWithTTL wires a configured JWT TTL (config.JWTTTL); NewServer
// keeps the historical 24h default for tests and existing callers.
func NewServerWithTTL(st *store.Store, jwt *auth.JWT, adminUser, version string, ttl time.Duration, log *slog.Logger) *Server {
	if ttl <= 0 || ttl > 24*time.Hour {
		ttl = 24 * time.Hour
	}
	s := &Server{Store: st, JWT: jwt, TTL: ttl, AdminUser: adminUser, Version: version, Log: log,
		logins:       newLoginLimiter(),
		smsFails:     map[string]*failSlot{},
		defaultTrust: parseCIDRs(defaultTrustCIDRs)}
	if st != nil {
		s.Epochs = st
	}
	return s
}

// Routes builds the handler tree:
//   - exact public endpoints: /api/v1/health, /api/v1/auth/login
//   - everything under /api/v1/ requires a valid JWT
func (s *Server) Routes() http.Handler {
	protected := http.NewServeMux()
	protected.HandleFunc("GET /api/v1/auth/me", s.handleMe)
	protected.HandleFunc("POST /api/v1/auth/logout", s.handleLogout)
	protected.HandleFunc("PUT /api/v1/auth/password", s.handlePasswordChange)
	protected.HandleFunc("GET /api/v1/inventory", s.handleInventory)
	protected.HandleFunc("GET /api/v1/inventory/categories", s.handleInventoryCategories)
	protected.HandleFunc("GET /api/v1/listings", s.handleListings)
	protected.HandleFunc("GET /api/v1/orders", s.handleOrders)
	protected.HandleFunc("GET /api/v1/templates", s.handleTemplates)
	protected.HandleFunc("GET /api/v1/templates/catalog", s.handleTemplateCatalog)
	protected.HandleFunc("PUT /api/v1/templates/blacklist", s.handleTemplateBlacklist)
	protected.HandleFunc("PUT /api/v1/inventory/{channel}/{asset_id}/cost", s.handleSetCost)
	protected.HandleFunc("GET /api/v1/strategies", s.handleStrategiesList)
	protected.HandleFunc("PUT /api/v1/strategies/global", s.handleStrategyUpdateGlobal)
	protected.HandleFunc("POST /api/v1/strategies/template", s.handleTemplateStrategyUpsert)
	protected.HandleFunc("DELETE /api/v1/strategies/template/{id}", s.handleTemplateStrategyDelete)
	protected.HandleFunc("GET /api/v1/audit", s.handleAuditList)
	protected.HandleFunc("GET /api/v1/price-actions", s.handlePriceActions)
	protected.HandleFunc("GET /api/v1/execution/status", s.handleExecutionStatus)
	protected.HandleFunc("GET /api/v1/dashboard", s.handleDashboard)
	protected.HandleFunc("GET /api/v1/jobs", s.handleJobsList)
	protected.HandleFunc("POST /api/v1/jobs/{name}/run", s.handleJobTrigger)
	protected.HandleFunc("GET /api/v1/channels", s.handleChannelsStatus)
	protected.HandleFunc("POST /api/v1/channels/uu/sms", s.handleUUSms)
	protected.HandleFunc("POST /api/v1/channels/uu/sms-verify", s.handleUUSmsVerify)
	protected.HandleFunc("PUT /api/v1/channels/uu", s.handleUUToken)
	protected.HandleFunc("PUT /api/v1/channels/eco", s.handleECOCreds)
	protected.HandleFunc("PUT /api/v1/channels/steam", s.handleSteamCreds)

	root := http.NewServeMux()
	root.HandleFunc("GET /api/v1/health", s.handleHealth)
	root.HandleFunc("POST /api/v1/auth/login", s.handleLogin)
	root.Handle("/api/v1/", s.requireAuth(protected))
	root.HandleFunc("/", func(w http.ResponseWriter, _ *http.Request) {
		writeErr(w, http.StatusNotFound, "not_found", "unknown endpoint")
	})
	return withRecover(s.Log)(withRequestLog(s.Log)(withSecurityHeaders(withBodyLimit(root))))
}

// maxBodyBytes caps request bodies — the largest legitimate payload is a small
// credentials JSON. Without this, the unauthenticated login endpoint doubles
// as an unbounded-decode memory amplifier.
const maxBodyBytes = 1 << 20 // 1 MiB

func withBodyLimit(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if r.Body != nil {
			r.Body = http.MaxBytesReader(w, r.Body, maxBodyBytes)
		}
		next.ServeHTTP(w, r)
	})
}

// withSecurityHeaders sets baseline hardening headers at the API layer so a
// direct :8080 connection (bypassing Caddy, e.g. LAN/debug access) is not
// header-naked. HSTS stays Caddy-only: it is meaningless (and harmful) over
// plain HTTP. Production must still go through Caddy (TLS + full CSP).
func withSecurityHeaders(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		h := w.Header()
		h.Set("X-Content-Type-Options", "nosniff")
		h.Set("X-Frame-Options", "DENY")
		h.Set("Referrer-Policy", "strict-origin-when-cross-origin")
		next.ServeHTTP(w, r)
	})
}

// ---- client identity (trusted-proxy aware) ----

// defaultTrustCIDRs: private/loopback peers may set X-Real-IP. In the
// documented deployment the backend port is never published — only the Caddy
// container (a private-range peer) can connect — so spoofing requires
// compromising the proxy itself. Direct public exposure must tighten this via
// TRUST_PROXY_CIDRS (e.g. to loopback only).
var defaultTrustCIDRs = []string{
	"127.0.0.0/8", "10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16",
	"::1/128", "fc00::/7",
}

// SetTrustProxies overrides the trusted peer CIDR list (from TRUST_PROXY_CIDRS).
// Must be called before serving; the list is parsed here so request-time
// trustList() reads are race-free.
func (s *Server) SetTrustProxies(cidrs []string) {
	parsed := parseCIDRs(cidrs)
	s.trustMu.Lock()
	defer s.trustMu.Unlock()
	s.trustProxies = parsed
}

func parseCIDRs(list []string) []*net.IPNet {
	out := make([]*net.IPNet, 0, len(list))
	for _, c := range list {
		if _, ipnet, err := net.ParseCIDR(c); err == nil {
			out = append(out, ipnet)
		}
	}
	return out
}

func (s *Server) trustList() []*net.IPNet {
	s.trustMu.RLock()
	defer s.trustMu.RUnlock()
	if len(s.trustProxies) > 0 {
		return s.trustProxies
	}
	return s.defaultTrust
}

// clientIP returns the rate-limit/audit identity: X-Real-IP only when the
// transport peer is a trusted proxy, else the peer address itself.
func (s *Server) clientIP(r *http.Request) string {
	host, _, err := net.SplitHostPort(r.RemoteAddr)
	if err != nil {
		host = r.RemoteAddr
	}
	if v := r.Header.Get("X-Real-IP"); v != "" {
		ip := net.ParseIP(host)
		for _, n := range s.trustList() {
			if ip != nil && n.Contains(ip) {
				return v
			}
		}
	}
	return host
}

func (s *Server) handleMe(w http.ResponseWriter, r *http.Request) {
	writeJSON(w, http.StatusOK, map[string]string{"username": userFrom(r.Context())})
}

// ---- helpers shared by handlers ----

func userFrom(ctx context.Context) string {
	v, _ := ctx.Value(ctxUser).(string)
	return v
}

func auditEntry(r *http.Request, action string, detail map[string]any) domain.AuditEntry {
	return domain.AuditEntry{
		Time:   time.Now().UTC(),
		Actor:  "user:" + userFrom(r.Context()),
		Action: action,
		Detail: detail,
	}
}

// audit writes an entry best-effort; nil store (tests) and write errors are non-fatal.
// The request context is detached: a browser disconnect mid-write must not
// silently drop the audit trail of a real write operation.
func (s *Server) audit(r *http.Request, action string, detail map[string]any) {
	if s.Store == nil {
		return
	}
	ctx, cancel := context.WithTimeout(context.WithoutCancel(r.Context()), 5*time.Second)
	defer cancel()
	if err := s.Store.InsertAudit(ctx, auditEntry(r, action, detail)); err != nil {
		s.Log.Warn("audit insert failed", "action", action, "err", err)
	}
}

// ---- middleware ----

type ctxKeyType int

const ctxUser ctxKeyType = 1

func (s *Server) requireAuth(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		h := r.Header.Get("Authorization")
		tok, ok := strings.CutPrefix(h, "Bearer ")
		if !ok || tok == "" {
			writeErr(w, http.StatusUnauthorized, "unauthorized", "missing bearer token")
			return
		}
		claims, err := s.JWT.Verify(tok)
		if err != nil {
			writeErr(w, http.StatusUnauthorized, "unauthorized", "invalid token")
			return
		}
		// Revocation (ADR-0006): tokens carry the session epoch they were
		// issued in; a bumped epoch (logout) invalidates every earlier token.
		epoch, err := s.sessionEpoch(r.Context())
		if err != nil {
			s.Log.Error("session epoch read failed", "err", err)
			writeErr(w, http.StatusUnauthorized, "unauthorized", "session unavailable")
			return
		}
		if claims.Ver != epoch {
			writeErr(w, http.StatusUnauthorized, "unauthorized", "token revoked")
			return
		}
		ctx := context.WithValue(r.Context(), ctxUser, claims.Sub)
		next.ServeHTTP(w, r.WithContext(ctx))
	})
}

func (s *Server) sessionEpoch(ctx context.Context) (int64, error) {
	if s.Epochs == nil {
		return 0, auth.ErrStoreUnavailable
	}
	epoch, err := s.Epochs.SessionEpoch(ctx)
	if err != nil {
		return 0, auth.FailClosedError(err)
	}
	return epoch, nil
}

// bumpSessionEpoch invalidates every outstanding token by advancing the epoch;
// returns the new epoch.
func (s *Server) bumpSessionEpoch(ctx context.Context) (int64, error) {
	if s.Epochs == nil {
		return 0, auth.ErrStoreUnavailable
	}
	return s.Epochs.BumpSessionEpoch(ctx)
}

func withRecover(log *slog.Logger) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			defer func() {
				if rec := recover(); rec != nil {
					log.Error("panic recovered", "path", r.URL.Path, "panic", rec)
					writeErr(w, http.StatusInternalServerError, "internal", "panic")
				}
			}()
			next.ServeHTTP(w, r)
		})
	}
}

func withRequestLog(log *slog.Logger) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			start := time.Now()
			next.ServeHTTP(w, r)
			log.Debug("http", "method", r.Method, "path", r.URL.Path, "dur_ms", time.Since(start).Milliseconds())
		})
	}
}