# docs/knowledge/spec/openapi.yaml
openapi: "3.0.3"
info:
  title: rent-auto API
  version: "0.10.0"
  description: |
    UU/ECO 双渠道 Steam 饰品租赁自动化系统 API。
    约定：
    - 前缀 /api/v1；除 health/login 外全部需要 Bearer JWT
    - 分页：query page(1-based) & page_size(<=200，round10 统一口径)；响应 {items,total}
    - 错误体：{code,message}；金额两位小数
    - 401 body code 恒为 `unauthorized`：前端据此清 token 回登录页
      （含登出吊销后的旧 token —— 会话纪元机制见 ADR-0006）
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: 登录获取 JWT（公开）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: token（绑定当前会话纪元）
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Err401" }
        "429": { $ref: "#/components/responses/Err429" }
...
  /auth/password:
    put:
      summary: 修改管理员密码并吊销全部既有会话
      description: >-
        校验当前密码后，单事务更新 bcrypt 密码与会话纪元；成功后当前 JWT 也失效，
        客户端清除本地 token 并重新登录。ADMIN_PASSWORD_HASH 外部托管时不允许通过此接口改密。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password, writeOnly: true }
                new_password:
                  type: string
                  format: password
                  writeOnly: true
                  description: UTF-8 编码后长度必须为 12–72 字节，不按字符数计
      responses:
        "200":
          description: 密码已更新，全部旧 JWT 已吊销
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        "400":
          description: bad_request（密码长度/JSON 非法）或 bad_credentials（当前密码错误）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Err401" }
        "409":
          description: password_managed_externally（外部托管）或 password_changed（并发改密冲突）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/Err429" }
        "500":
          description: 内部错误，错误信息脱敏；不承诺密码变更成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /inventory:
    get:
      summary: 库存列表（分页）
      parameters:
        - { name: channel, in: query, schema: { type: string, enum: [uu, eco] } }
        - { name: status, in: query, schema: { type: string, enum: [in_stock, listed, leased, locked, sold, missing] } }
        - { name: search, in: query, schema: { type: string }, description: 名称或资产 ID 模糊匹配 }
        - { name: category, in: query, schema: { type: string }, description: 品类精确匹配；空品类以未分类表示 }
        - { name: hash_name, in: query, schema: { type: string }, description: 模板精确匹配 }
        - { name: asset_id, in: query, schema: { type: string }, description: 资产精确匹配 }
        - { name: cost_missing, in: query, schema: { type: boolean }, description: true 仅返回未录入正数成本的记录 }
        - { name: sort, in: query, schema: { type: string, enum: [name, price_desc, price_asc, cost_desc] }, description: 缺省或未知值按入库 ID 升序；排序作用于完整集合 }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页库存
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/InventoryItem" }
  /listings:
    get:
      summary: 上架列表（分页）
      parameters:
        - { name: channel, in: query, schema: { type: string, enum: [uu, eco] } }
        - { name: state, in: query, schema: { type: string }, description: 实际状态过滤 }
        - { name: hash_name, in: query, schema: { type: string } }
        - { name: asset_id, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string }, description: 名称或资产 ID 或货架编号模糊匹配 }
        - { name: mismatch, in: query, schema: { type: boolean }, description: 仅状态不一致；含持续不一致标记 }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页上架
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/ListingRow" }
  /orders:
    get:
      summary: 租赁订单（分页）
      parameters:
        - { name: channel, in: query, schema: { type: string, enum: [uu, eco] } }
        - { name: status, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string }, description: 名称或订单号模糊匹配 }
        - { name: hash_name, in: query, schema: { type: string } }
        - { name: asset_id, in: query, schema: { type: string } }
        - { name: order_type, in: query, schema: { type: string, enum: [short, long, buyout] } }
        - { name: view, in: query, schema: { type: string, enum: [attention, active, history] }, description: attention=待支付/发货/归还/仲裁/违约/未知；active=leasing；history=done/bought_out/cancelled }
        - { name: sort, in: query, schema: { type: string, enum: [due_asc, started_desc, amount_desc] }, description: 缺省或未知值按 ID 降序；空日期最后 }
        - { name: since, in: query, schema: { type: string, format: date-time }, description: started_at 含下界 }
        - { name: until, in: query, schema: { type: string, format: date-time }, description: started_at 不含上界；须晚于 since，非法返回 400 }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页订单
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/OrderRow" }
...
  /templates/catalog:
    get:
      summary: 可搜索的分页模板目录
      parameters:
        - { name: search, in: query, schema: { type: string }, description: hash_name 或 display_name 模糊匹配 }
        - { name: blacklisted, in: query, schema: { type: boolean } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页目录与完整匹配数量
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items: { type: array, items: { $ref: "#/components/schemas/Template" } }
        "400": { $ref: "#/components/responses/Err400" }
  /execution/status:
    get:
      summary: 当前执行许可（只读；并不代表任务正在运行）
      description: 环境 DRY_RUN_DEFAULT、全局总闸及有效模板许可共同决定。缺少环境配置或读取失败时不返回猜测的模式。
      parameters:
        - { name: hash_name, in: query, schema: { type: string }, description: 省略时只检查全局；带值时检查模板有效策略 }
      responses:
        "200":
          description: 执行许可与限制原因
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExecutionStatus" }
        "503": { $ref: "#/components/responses/Err503" }
  /price-actions:
    get:
      summary: 完整历史定价动作（只读，保留实际模拟/结果/空值）
      description: 不以当前策略重建历史。无策略版本或行情快照的记录不虚构这些元数据。decision 递归脱敏，error 仅公开通用提示。
      parameters:
        - { name: channel, in: query, schema: { type: string, enum: [uu, eco] } }
        - { name: hash_name, in: query, schema: { type: string } }
        - { name: asset_id, in: query, schema: { type: string } }
        - { name: id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: listing_id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: action, in: query, schema: { type: string, enum: [publish, reprice, delist, skip] } }
        - { name: mode, in: query, schema: { type: string, enum: [dry_run, real] } }
        - { name: result, in: query, schema: { type: string, enum: [success, failed, skip] }, description: success 为非跳过且 success=true；failed 为 success=false；skip 为动作跳过 }
        - { name: since, in: query, schema: { type: string, format: date-time }, description: ts 含下界 }
        - { name: until, in: query, schema: { type: string, format: date-time }, description: ts 不含上界；须晚于 since }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 按 id 降序的分页记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items: { type: array, items: { $ref: "#/components/schemas/PriceAction" } }
        "400": { $ref: "#/components/responses/Err400" }
  /templates:
    get:
      summary: 模板清单（含锚点/黑名单标记）
      description: >-
        分页可选：page_size 缺省或 ≤0 时返回全量目录并忽略 page（面板黑名单表与模板选择器
        依赖完整清单）；page_size>0 时按 1-based page 切页，先将每页数量钳至 200 再计算偏移。
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 0, maximum: 200 },
            description: "≤0 或缺省 = 全量" }
      responses:
        "200":
          description: 模板数组
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Template" }
  /templates/blacklist:
    put:
      summary: 切换模板拉黑状态（退出路由与锚点合成；审计 template.blacklist）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hash_name]
              properties:
                hash_name: { type: string }
                blacklisted: { type: boolean }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "404": { $ref: "#/components/responses/Err400" }
  /inventory/{channel}/{asset_id}/cost:
    put:
      summary: 录入/修正成本价
      parameters:
        - { name: channel, in: path, required: true, schema: { type: string, enum: [uu, eco] } }
        - { name: asset_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cost]
              properties:
                cost: { type: number, format: double, minimum: 0 }
      responses:
        "200":
          description: 回显成本
          content:
            application/json:
              schema:
                type: object
                properties:
                  cost: { type: number, format: double }
        "400": { $ref: "#/components/responses/Err400" }
...
  /strategies/template:
    post:
      summary: 创建/替换模板级覆盖策略（每模板一行；新建行默认 dry-run）
      description: >-
        params 是稀疏覆盖对象；保存会替换模板的原覆盖对象，未出现的参数继续继承当前全局值，
        空对象清除全部参数覆盖。模板与全局写入共享事务锁，先校验当前全局与本次覆盖的有效组合，
        非法时 400 且整次修改不生效。面板编辑仅提交明确覆盖的参数，并保留执行开关与优先级。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hash_name, channel_route]
              properties:
                hash_name: { type: string }
                channel_route: { type: string, enum: [uu_only, eco_only, both, uu_primary_eco_fallback] }
                params:
                  type: object
                  description: 分组结构（baseline/factor/guardrails/uu_max_days/eco_max_days），深覆盖全局
                real_execution_enabled:
                  type: boolean
                  description: 新建时缺省=false；更新时缺省=保留原值。模板 true 不能绕过全局真实执行总闸或 DRY_RUN_DEFAULT
                priority: { type: integer, format: int32, default: 0, description: 本次写入的优先级；更新时应显式传回原值以保留 }
      responses:
        "200":
          description: 行 id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/Err400" }
  /strategies/template/{id}:
    delete:
      summary: 删除模板级覆盖（立即回落全局策略）
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
        "404": { $ref: "#/components/responses/Err400" }
  /strategies/global:
    put:
      summary: 更新全局策略（参数/路由/真实执行开关，单事务生效）
      description: >-
        修改 params 或开启真实执行时，在与模板写入共享的事务锁内校验全局参数及所有已启用模板的
        有效继承组合。任一组合非法则返回 400，参数/路由/执行标志全部不变。
        仅关闭全局真实执行不依赖参数合法性，保证存量错误配置仍可停机。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                params: { type: object, description: 策略参数 JSON（深合并默认值） }
                route: { type: string, enum: [uu_only, eco_only, both, uu_primary_eco_fallback],
                  description: 路由（别名；与 channel_route 二选一，channel_route 优先） }
                channel_route: { type: string, enum: [uu_only, eco_only, both, uu_primary_eco_fallback],
                  description: 路由（面板 canonical 字段；与 route 双收） }
                real_execution_enabled: { type: boolean, description: 全局真实执行总闸；false 强制全部模板 dry-run，true 仍受模板开关与 DRY_RUN_DEFAULT 约束 }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
        "400": { $ref: "#/components/responses/Err400" }
  /audit:
    get:
      summary: 审计日志（最近 N 条）
      description: page_size 缺省或 ≤0 使用 50，超过 200 钳至 200；page 偏移按规范化后的每页数量计算。detail 递归脱敏，不截断可公开的长详情。时间严格按 RFC3339 的 [since,until)；非法或倒置区间返回 400。
      parameters:
        - { name: action, in: query, schema: { type: string }, description: 动作名称精确匹配 }
        - { name: channel, in: query, schema: { type: string, enum: [uu, eco] } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: until, in: query, schema: { type: string, format: date-time } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 分页审计
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/AuditEntry" }
...
  /jobs:
    get:
      summary: 最近任务状态列表（不提供逐次运行历史或与某次点击的关联）
      responses:
        "200":
          description: 任务数组
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/JobStatus" }
        "503": { $ref: "#/components/responses/Err503" }
  /jobs/{name}/run:
    post:
      summary: 手动触发一次任务（立即执行，10 分钟预算，停机前会被等待）
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 已受理，不能视为执行完成；reprice 始终是全局任务，不接受列表筛选
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
        "400":
          description: 任务未知或正在运行
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /channels/uu/sms:
    post:
      summary: 发起 UU 短信登录（三种模式：下行验证码 / 上行短信指引 / 图形验证码）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
                session_id: { type: string, description: 重试时回传上次返回值 }
                captcha:
                  type: object
                  description: captcha 模式重试时携带 TCaptcha 结果；必须同时回传首次响应的 session_id（票据链与 session 绑定），否则 400
                  properties:
                    ticket: { type: string }
                    randstr: { type: string }
                    req_ticket: { type: string }
      responses:
        "200":
          description: 发送结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UUSmsResponse" }
        "400": { $ref: "#/components/responses/Err400" }
        "502": { $ref: "#/components/responses/Err502" }
  /channels/uu/sms-verify:
    post:
      summary: 提交短信验证码完成登录并保存凭证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, phone, code]
              properties:
                session_id: { type: string }
                phone: { type: string }
                code: { type: string }
                login_req_ticket: { type: string, description: 下行模式成功响应携带的关联票据 }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
        "400":
          description: 验证码错误等上游拒绝（非面板会话失效）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /channels/uu:
    put:
      summary: 手动导入 UU Token（5050 登录门禁的替代路径）
      description: |
        平台限制第三方客户端短信登录（code=5050）后，管理员在官网/APP 登录
        获取 JWT token，粘贴导入。服务端先以 getUserInfo 验证 token 有效性，
        验证通过后 AES-256-GCM 加密落库并即时重建 UU 适配器。审计仅记
        token 尾 8 位指纹。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: UU JWT（三段式）}
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  fingerprint: { type: string, description: UU token 尾 8 位 }
        "400":
          description: token 为空或上游校验失败（84101 失效/网络错误）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /channels/eco:
    put:
      summary: 保存 ECO 凭证（加密落库并即时重建适配器）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partner_id, private_key_pem]
              properties:
                partner_id: { type: string }
                private_key_pem: { type: string, description: PKCS8 PEM }
                steam_id: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  fingerprint: { type: string, description: 私钥 SHA-256 前 12 位 }
        "400": { $ref: "#/components/responses/Err400" }
  /channels/steam:
    put:
      summary: 保存 Steam 凭证（触发真实登录与会话持久化）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, shared_secret, identity_secret]
              properties:
                username: { type: string }
                password: { type: string }
                shared_secret: { type: string }
                identity_secret: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  fingerprint: { type: string, description: Guard shared_secret SHA-256 前 12 位 }
        "400": { $ref: "#/components/responses/Err400" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
      description: 缺省或 ≤0 使用 50，超过 200 钳至 200；偏移按规范化后的数量计算
  responses:
    Err503:
      description: 状态或服务未初始化；客户端显示未知并允许重试
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Err400:
      description: 参数错误
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Err401:
      description: 未认证/已吊销
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Err429:
      description: 登录尝试过多（固定窗口锁定）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Err502:
      description: 上游平台失败
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    PagedEnvelope:
      type: object
      properties:
        total: { type: integer }
...
    InventoryItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        channel: { $ref: "#/components/schemas/Channel" }
        asset_id: { type: string }
        hash_name: { type: string }
        market_hash_name: { type: string }
        template_id: { type: integer, format: int64, nullable: true }
        mark_price: { type: number, format: double }
        tradable: { type: boolean }
        status:
          type: string
          enum: [in_stock, listed, leased, locked, sold, missing]
          description: missing 表示完整库存快照中未再出现，不等于已售出
        cost_basis: { type: number, format: double, description: "成本价（PUT /inventory/{channel}/{asset_id}/cost 录入）" }
        category: { type: string }
        cost_source: { type: string }
        cost_modified_at: { type: string, format: date-time, nullable: true }
        last_synced_at: { type: string, format: date-time, nullable: true }
...